+++ client/src/components/Languages/Language.jsx
@@ -32,9 +32,9 @@ export const languageList = [
   },
 ]
 
+export const languages = ['en', 'nl', 'fr', 'pt'];
 
-
-export function Language() {
+export default function Language() {
   const currentLanguageCode = cookies.get('i18next') || 'en'
   const currentLanguage = languageList.find((l) => l.code === currentLanguageCode)
   const { t } = useTranslation()
@@ -46,43 +46,43 @@ export function Language() {
     document.title = t('app_title')
   }, [currentLanguage, t])
 
+  const onClickChangeLanguage = (code) => () => {
+    i18next.changeLanguage(code);
+  }
+
   return (
-    <>
-      <div className="language">
-        <div className="dropdown">
-          <button
-            className="btn btn-link dropdown-toggle"
-            type="button"
-            id="dropdownMenuButton1"
-            data-bs-toggle="dropdown"
-            aria-expanded="false"
-          >
-            <img className='globeIcon' src={GlobeIcon} alt="globeIcon" />
-          </button>
-          <ul className="dropdown-menu" aria-labelledby="dropdownMenuButton1">
-            {languageList.map(({ code, name, country_code }) => (
-              <li key={country_code}>
-                <button
-                  className={classNames('dropdown-item', {
-                    disabled: currentLanguageCode === code,
-                  })}
-                  onClick={() => {
-                    i18next.changeLanguage(code)
+    <div className="language">
+      <div className="dropdown">
+        <button
+          className="btn btn-link dropdown-toggle"
+          type="button"
+          id="dropdownMenuButton1"
+          data-bs-toggle="dropdown"
+          aria-expanded="false"
+        >
+          <img className='globeIcon' src={GlobeIcon} alt="globeIcon" />
+        </button>
+        <ul className="dropdown-menu" aria-labelledby="dropdownMenuButton1">
+          {languageList.map(({ code, name, country_code }) => (
+            <li key={country_code}>
+              <button
+                className={classNames('dropdown-item', {
+                  disabled: currentLanguageCode === code,
+                })}
+                onClick={onClickChangeLanguage(code)}
+              >
+                <span
+                  className={`flag-icon flag-icon-${country_code} mx-2`}
+                  style={{
+                    opacity: currentLanguageCode === code ? 0.5 : 1,
                   }}
-                >
-                  <span
-                    className={`flag-icon flag-icon-${country_code} mx-2`}
-                    style={{
-                      opacity: currentLanguageCode === code ? 0.5 : 1,
-                    }}
-                  ></span>
-                  {name}
-                </button>
-              </li>
-            ))}
-          </ul>
-        </div>
+                ></span>
+                {name}
+              </button>
+            </li>
+          ))}
+        </ul>
       </div>
-    </>
+    </div>
   )
 }
